fix: allow bidirectional raw conversion due to switch away from interface (#33)
* fix: allow bidirectional raw conversion

This allows for bidirectional raw conversion by adding a path in convert that specifically converts back into Raw.

This is especially useful now because we switched from using interface types in our structs to only raw, so it is important that our conversion works as intended everytime.

```
// Convert transforms one Typed object into another. Both 'from' and 'into' must be non-nil pointers.
//
// Special Cases:
//   - Raw → Raw: performs a deep copy of the underlying []byte data.
//   - Raw → Typed: unmarshals YAML from raw.Data into the typed object (if the type is registered).
//   - Typed → Raw: marshals the object to JSON, applies canonicalization, and stores the result in Raw.Data. (See Raw.UnmarshalJSON)
//   - Typed → Typed: performs a deep copy using Typed.DeepCopyTyped, with reflection-based assignment.
//
// Errors are returned if:
//   - Either argument is nil.
//   - A type is not registered in the Scheme (for Raw conversions).
//   - A reflection-based assignment fails due to type mismatch.
```

* fix: introduce type defaulting in runtime

This allows to convert between types even when the type attribute was not set, similar to kubernetes client runtime.

It also adds a type introspection in raw that is able to set the type from a raw json in case it wasnt set before

* feat: switch to SetType instead of json typing

* feat: switch to SetType instead of json typing

* feat: switch to SetType instead of json typing