fix: move non oci-compliant media types into layers instead of single layer artifacts (#77)
when storing local blobs that are not a valid OCI compliant media type such as oci image manifests or indexes, it is a better approach to store them as layers instead of single layer OCI artifacts. While I originally tried this there are distinct disadvantages:

1. It is introducing overhead of one manifest per local blob which can escalate requests
2. It breaks with existing OCI implementations that rely on local blobs being present for all media types, not just OCI specific ones
3. It makes it so that the localBlob specification has a mediaType set that correlates to the single layer oci artifact instead of the actual layer, which would result in either a new media type or other shenanigans to user input we want to avoid.

So lets move this to layers instead. Fortunately the layer storage already exists.