fix: implemented repeated read through ReadSeeker interface and automatic seek back (#223)
* fix: implemented repeated read through seek

this implements a default read / seek on the buffered reader so that everytime the buffer has been fully read, it gets reset fully and can be read from again.

This allows the following flow:

1. Read data from source (e.g. file) in memory
2. Read from memory after first read from file
3. Read again from memory henceforth (this was previously not possible)

* fix: implemented repeated read through seek

this implements a default read / seek on the buffered reader so that everytime the buffer has been fully read, it gets reset fully and can be read from again.

This allows the following flow:

1. Read data from source (e.g. file) in memory
2. Read from memory after first read from file
3. Read again from memory henceforth (this was previously not possible)