fix!: constructor fixes stemming from `add cv` integration work (#219)
* feat: rework component constructor conversions

this unifies and standardizes functions for constructor conversion from and to the constructor runtime, the constructor specification and the descriptor runtime

* feat: rework constructor to allow replacement policy

previously it was not possible to include replacement strategies for component constructors. This has now been included with a flag.

Also the redundant validation was removed as there was a validation through specification already.

in the default scheme there were some left-overs that were removed as well.

For the resource repository, a credential can now be passed as previously there was no way to pass in credentials.

For local blobs, a standard defaulting to application/octet-stream has been included so that new local blobs without declared media type have a set generic media type

* feat: rework component constructor conversions

this unifies and standardizes functions for constructor conversion from and to the constructor runtime, the constructor specification and the descriptor runtime

* chore(test): add missing policy tests

* chore(docs): add small doc line for resource rpeository

* chore(lint): fixup

* feat: add credential support for digest processor

* chore: bump

* chore: cleanup conflict strategy processing

* chore: fixup