fix: proper tag listing (#277)
* fix: proper tag listing

previously the lib only used the tags resolve functionality when listing tags. this could cause issues if there were other tagged manifests that were not ocm component versions. also the tagging was not properly concurrent but worked in concurrent blocks equivalent to the tag list page size.

now tag listing works concurrently based on the resolution of the manifest and then the interpretation of the component version annotation on the manifest to differentiate OCM and non OCM versions. While less efficient (a fetch call more per tag), due to the efficiency gains during listing, we should still be good to go.

* chore: fixup

Co-authored-by: Fabian Burth <fabian.burth@sap.com>

---------

Co-authored-by: Fabian Burth <fabian.burth@sap.com>