chore(docs): ocm resource download ADR (#364)
* chore(docs): document resource downloader

Signed-off-by: Jakob Möller <jakob.moeller@sap.com>

* chore(docs): document resource downloader

Signed-off-by: Jakob Möller <jakob.moeller@sap.com>

* chore(docs): document resource downloader

Signed-off-by: Jakob Möller <jakob.moeller@sap.com>

* feat(oci)!: introspect OCI Layouts with graph

This makes it so that the OCI library can now dynamically discover top level artifacts in OCI layouts passed to upload scenarios. This is achieved by building a successor graph dynamically based on the index. The call itself is implemented to be extremely efficient for 90% of our cases. (1-10 main artifacts with <100 references).

This effectively allows us to use layouts without passing in a source access and we can simplify our OCI repository interfaces signficantly. This means we are now much closer to the old OCMv1 Main Artifact Fetching in Artifact Sets.

Signed-off-by: Jakob Möller <jakob.moeller@sap.com>

* docs: adjust adr via PR

Signed-off-by: Jakob Möller <jakob.moeller@sap.com>

---------

Signed-off-by: Jakob Möller <jakob.moeller@sap.com>