chore(ci): add OpenSSF Scorecard Supply Chain Results (#411)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it

Scorecard is an automated tool that assesses a number of important
heuristics
[("checks")](https://github.com/ossf/scorecard?tab=readme-ov-file#scorecard-checks)
associated with software security and assigns each check a score of
0-10. You can use these scores to understand specific areas to improve
in order to strengthen the security posture of your project. You can
also assess the risks that dependencies introduce, and make informed
decisions about accepting these risks, evaluating alternative solutions,
or working with the maintainers to make improvements.

#### Which issue(s) this PR fixes
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->

Signed-off-by: Jakob Möller <jakob.moeller@sap.com>