fix(oci): adjust oci to be able to deal ocm v1 oci manifests of ocm components (#419)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it
In ocm v1, the version - even if given - is omitted from the identity if
the identity is already unique without the version.
In ocm v2, the version is always part of the identity.

Within oci, the identity of resources and sources is written as an layer
annotation. ocm v2 leverages this annotation to filter for the correct
layer(s) of a resource. Due to above behavior, in case the version is
given but is not required for the identity to be unique, the identities
were not **equal** and thus, the layer were not found.

To fix this, this PR exchanges the **equality** matcher with a
**subset** matcher. Since `(identity) is subset of (identity +
version)`, the layers can now be found.

#### Which issue(s) this PR fixes
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->

---------

Signed-off-by: Fabian Burth <fabian.burth@sap.com>