feat(oci): use interfaces from repositories module (#421)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it
Related to:
https://github.com/open-component-model/open-component-model/pull/403

This PR added the ocm repository interfaces to a module that is intended
to only work on the abstraction level of ocm repositories (and
therefore, does not have storage technology specific code) and can be
imported by oci.

#### Which issue(s) this PR fixes
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->

---------

Signed-off-by: Fabian Burth <fabian.burth@sap.com>