feat(construct): update interfaces and omit empty name and/or version in constructor `ToIdentity` (#428)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it
This PR updates the constructor module to use the interfaces from the
central repositories module instead of from the oci module. Besides, it
also aligns the constructors `ToIdentity` handling with the descriptors
`ToIdentity` handling, where name and/or. version are omitted if empty.

#### Which issue(s) this PR fixes
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->

---------

Signed-off-by: Fabian Burth <fabian.burth@sap.com>
Co-authored-by: Jakob Möller <jakob.moeller@sap.com>