chore(ci): make CLI CTF build faster by caching the CLI go dependencies (#444)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it

this makes it so that the CTF build now uses cached values from the CLI
go.mod so that it becomes faster and we dont have to maintain a separate
go.mod. It does this by checking out the CLI first

#### Which issue(s) this PR fixes
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->

Signed-off-by: Jakob Möller <jakob.moeller@sap.com>
Co-authored-by: Fabian Burth <fabian.burth@sap.com>