feat: add `dir` input to cli (#450)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it

This PR adds the `dir` input method to the CLI. I.e. when merged, it
will be possible to do `ocm add cv` command with `--constructor` flag
pointing to a `component-constructor.yaml`, which has a resource with an
`input` of type `dir`.

Supported scope:
* Absolute paths or paths relative to the working directory in the
`path` filed
* Uncompressed input (`compress` field omitted or set to false)

Out of scope:
* Paths relative to component constructor
(open-component-model/ocm-project#565)
* `compress: true` (open-component-model/ocm-project#566)

#### Which issue(s) this PR fixes
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->

Signed-off-by: Ilya Khandamirov <ilya.khandamirov@sap.com>