fix: optimize layout copying (#501)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it

This changes the underlying CopyToOCILayoutInMemory implementation so
that any call to it triggers a layout copy asynchronously. This omits
the buffer and only causes buffering at the usage state if the digest or
size are requested.

This effectively makes it so that the blobs returned only compute size
and digest when needed, not by default

#### Which issue(s) this PR fixes
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->

Signed-off-by: Jakob Möller <jakob.moeller@sap.com>
Co-authored-by: ikhandamirov <108289993+ikhandamirov@users.noreply.github.com>