chore!: simplify and refactor blob caching (#513)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it

when going through the blob package I have noticed that the existing
cache implementation is quite complex as it loads data dynamically in
every method. by changing the construction of the cache to happen
eagerly in the method call we can strongly simplify the usage of the
cache (one just has to call `Cache(blob)` now and it will actually do
the caching), and clean up existing code

#### Which issue(s) this PR fixes
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->

We dont actually have any usages of this Package yet and thats also
because we didnt need blob based caching before. now we do and it turns
out we rather tend to use inmemory Blob with Load.

I would like to slowly but surely get rid of inmemory Blob and replace
it with a direct blob that is cached. That way we do not have any
strange inmemory default behaviors anymore

Signed-off-by: Jakob Möller <jakob.moeller@sap.com>
Co-authored-by: ikhandamirov <108289993+ikhandamirov@users.noreply.github.com>