fix: support `compress` option for `file` and `dir` inputs (#512)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it

This PR makes it possible to create resource from `file`and `dir` with
`Compress` option set to `true`. In such cases, when creating respective
OCI layer, the size and the digest of the blob are now calculated
dynamically.

The use of the fix is not limited to `file`and `dir` inputs, but
applicable to all cases, when the size or the digest of the blob are not
known in advance.

#### Which issue(s) this PR fixes
Fixes open-component-model/ocm-project#566

---------

Signed-off-by: Ilya Khandamirov <ilya.khandamirov@sap.com>
Co-authored-by: Jakob Möller <jakob.moeller@sap.com>