feat: direct blob (#503)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it

this implements a direct blob, a blob that can be read from many times
independently but that does not buffer itself backed by `io.Reader`

this is especially useful when forwarding already existing buffers such
as bytes.Buffer or plain []byte slices, or using readers created by
`strings.NewReader` or `bytes.NewReader`. All of these implementations
are fully compatible with this implementation. Even using an os file
descriptors or http responses directly is valid.

#### Which issue(s) this PR fixes
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->

One caveat towards the existing "inmemory" implementation is that
because it does not buffer data itself, we cannot know the digest in
advance and if someone wants to compute it, they will have to buffer the
data themselves and compute it dynamically. The obvious upside though is
that we can write blob providers that do not implicitly cache data on
the source, but rather get cached on demand.

part of https://github.com/open-component-model/ocm-project/issues/580

---------

Signed-off-by: Jakob Möller <jakob.moeller@sap.com>