feat: adjust filename based on oci artifact media type (#535)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it

This is part of
https://github.com/open-component-model/ocm-project/issues/553 effort.
As documented in the code, unfortunately, packaging and creating files
for HELM MUST follow a certain requirement detailed here:
https://helm.sh/docs/topics/charts/#charts-and-versioning.

We could do several other things like we initially planned with CEL
integration, but this is an easier path for now to go.

#### Which issue(s) this PR fixes
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->

---------

Signed-off-by: Gergely Brautigam <gergely.brautigam@sap.com>