feat: support stateful, serialized io.Reader's in direct blobs (#541)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it

when we don't have io.ReaderAt backed implementations of blobs (for
example when working with io.Pipe which is not resettable), then we need
to base our implementation solely on io.Reader. However the blobs
ReadCloser() methods can in theory be concurrently called, so we need a
serializing blob that is able to protect from bad reads.

#### Which issue(s) this PR fixes
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->
This fixes not being able to use pure io.Readers and allows direct blobs
to receive an io.Reader natively, switching to serialized access under
the hood if needed.

Needed to actually migrate input methods like dir
(https://github.com/open-component-model/ocm-project/issues/580)

---------

Signed-off-by: Jakob Möller <jakob.moeller@sap.com>