fix: wrap layer data with `io.NopCloser` to prevent premature closing by storage implementation (#567)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it

the ORAS OCI implementation opinionates push by automatically closing a
ReadCloser. since we also close, we would receive a ErrClosed which we
want to avoid when working with OCI. That's why this should be wrapped
with io.NopCloser

#### Which issue(s) this PR fixes
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->

part of the helm input binary test did with @Skarlso
part of https://github.com/open-component-model/ocm-project/issues/490

Signed-off-by: Jakob Möller <jakob.moeller@sap.com>
Co-authored-by: Gergely Brautigam <gergely.brautigam@sap.com>