fix: add another conformance for legacy JSON normalization (v3) (#542)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it

Implemented a conformance testing suite for validating legacy JSON
normalization (v3). Included test datasets and adjusted handling of
provider parsing logic to handle both string and byte inputs to not
break conformance

#### Which issue(s) this PR fixes
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->

fixed a broken case in which the wrong normalisation was computed. If
the provider only had a single string value, it was incorrectly
normalised to the string instead of a map with the "name" key

this paves the way to get exactly the same output as with old OCM CLI
for the newest normalisation algorithm.

relates to
https://github.com/open-component-model/ocm-project/issues/579

---------

Signed-off-by: Jakob Möller <jakob.moeller@sap.com>
Co-authored-by: Fabian Burth <fabian.burth@sap.com>