fix: run plugin download integration separately as a go module (#581)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it

Fixes
https://github.com/open-component-model/open-component-model/issues/579

In order to avoid bloating the go mod file and make sure that the tests
aren't run as part of the normal unit test flow, we cordon it off into
its own module and mark give it a special Task file.

#### Which issue(s) this PR fixes
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->

---------

Signed-off-by: Gergely Brautigam <182850+Skarlso@users.noreply.github.com>