feat(cli): json/yaml graph renderer (#574)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it
This PR introduces a renderer for lists of objects, rendering them as
json or yaml lists:

```
[
  "A",
  "B",
  "C",
  "D"
]
```
```
- A
- B
- C
- D
```

**IMPORTANT:**
This implementation would be a breaking change to ocm v1, where json or
yaml lists were rendered as:
```
{
  "items": [
      "A",
      "B",
      "C",
      "D"
  ]
}
```
```
A
---
B
---
C
---
D
```

Considerations:
- the first one makes it to be parsable more intuitively
- the first one is the native json/yaml list marshaling

BUT
WILL DEFINITELY break workflows depending on the json or yaml output of
the current CLI (if there are any).

It renames the package from `renderer` to `render` as it also defines
general rendering utilities.
It defines `CursorUp` and `EraseLine` itself instead of using the one
from the library to make the `render` package independent of the
library.

#### Which issue(s) this PR fixes
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->
Contributes to
https://github.com/open-component-model/ocm-project/issues/563

---------

Signed-off-by: Fabian Burth <fabian.burth@sap.com>
Signed-off-by: Jakob Möller <jakob.moeller@sap.com>
Co-authored-by: Jakob Möller <jakob.moeller@sap.com>