feat: add a wrapper to blob to wrap any ReadOnlyBlob to io.ReadCloser (#593)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it

This creates a wrapper for ReadOnlyBlobs so they can act like a
MediaType aware blob.

This is part of
https://github.com/open-component-model/open-component-model/pull/573.

The reason here is, that we don't want filesystem to be modified to be
media type aware. Rather we can wrap it in a direct blob that will be
able to call the underlying reader and even add capabilities to it, like
being media type aware.

Such a wrapper could be created like this:
```go
w := direct.New(direct.NewReadCloserWrapper(b), direct.WithMediaType(location.MediaType))
```

This will create a wrapper that is a `blob.ReadOnlyBlob`.

#### Which issue(s) this PR fixes
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->

---------

Signed-off-by: Gergely Brautigam <182850+Skarlso@users.noreply.github.com>