feat(dag)!: concurrent traversal and interface for traversal func (#605)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it
We need the concurrent traversal to be able to traverse a dag from
multiple roots in the ocm get command, in case we have multiple
components in a `--recursive` call.
We introduced the interface because we agreed once that this makes it
easier to navigate the code. Besides, it provides flexibility.

#### Which issue(s) this PR fixes
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->
Contributes to:
https://github.com/open-component-model/ocm-project/issues/563

---------

Signed-off-by: Fabian Burth <fabian.burth@sap.com>