fix: manually update blob dependency to unblock monorepo job (#623)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it
Currently, the monorepo upgrade job is blocked because of a test failure
in the dir input methods. This PR resolves the test failure and manually
upgrades the blob dependency of the dir input method to work towards
unblocking the monorepo upgrade job.

This is the change in the blob module that broke the current behaviour:
```
- func (s *osFileSystem) ReadDir(name string) ([]fs.DirEntry, error) {
-	return os.ReadDir(filepath.Join(s.base, name))
+ func (s *RootFileSystem) ReadDir(name string) ([]fs.DirEntry, error) {
+	return s.root.FS().(fs.ReadDirFS).ReadDir(name)
```

With the original variant, `name == ""` resulted in a valid path due to
`filepath.Join(s.base, name)`. With the new variant, we call
`rootFS.ReadDir(name)` which internally calls `fs.ValidPath(name)`.
`fs.ValidPath(name)` has the following documentation:

> ...
> // Path names passed to open are UTF-8-encoded,
> // unrooted, slash-separated sequences of path elements, like “x/y/z”.
> // **Path names must not contain an element that is “.” or “..” or the
empty string,**
> // **except for the special case that the name "." may be used for the
root directory**.
> // Paths must not start or end with a slash: “/x” and “x/” are
invalid.
> ...

#### Which issue(s) this PR fixes
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->

Signed-off-by: Fabian Burth <fabian.burth@sap.com>