fix(ci): streamline docker build and push process for multi-architecture support (#698)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it

When we push our image, which is multi-arch, we cannot use a "loaded"
version from the docker daemon we use for e2e testing, because that
doesn't support multi-arch. That means we need to push directly from the
builder instance instead of loading it first.

#### Which issue(s) this PR fixes
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->
Fixes
https://github.com/open-component-model/open-component-model/actions/runs/17262857226/job/48988854136

Signed-off-by: Jakob Möller <jakob.moeller@sap.com>