feat(dag)!: introduce dag processing logic (#770)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it
This PR introduces logic to perform concurrent topological traversal of
a dag, with child nodes always being processed before its parent node.

- It also fixes a "bug" in the `Reverse` methods where attributes have
not been copied (it was simply not implemented).
- It renames `Traverse` to `Discover` since this caused a lot of
confusion.
- It fixes a flaw in the `Discover`/`DiscoverNeighbors` logic that
allowed the discover neighbors function to add attributes to the
neighbors which had the potential to cause races. Now, the discover
neighbors logic is only passed the vertex ID and can only return the
neighbor vertices IDs.
- This is sufficient for our component cases for now. If we need
attributes during discovery in the future, we'd have to add them to the
edges.

#### Which issue(s) this PR fixes
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->
Contributes to:
https://github.com/open-component-model/ocm-project/issues/564

---------

Signed-off-by: Fabian Burth <fabian.burth@sap.com>