feat(dag): dag based construct processing (#772)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it
The current construct implementation processes all the components in a
constructor concurrently. Since we want to be able to add the digest of
component references to the constructed component descriptors. To even
calculate the digest of the referenced component descriptor, that
descriptor has to be constructed already.

As a conclusion, we need a graph processing logic, where all children
are constructed before its parent is being constructed.

#### Which issue(s) this PR fixes
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->
https://github.com/open-component-model/ocm-project/issues/564

---------

Signed-off-by: Fabian Burth <fabian.burth@sap.com>