feat(plugin): introduce signing handler plugin support (#885)
#### What this does
- Added a new `SigningRegistry` to manage signing handler plugins.
- Integrated signing handler support into the plugin manager and
registries.
- Updated dependencies in `go.mod` and `go.sum` to include the signing
library.
- Adjusted test cases to align with the new signing plugin
functionality.

_Note: I have made it so that Signing and Verification expect the same
Config Type right now. Technically this is a bit of a conflict with the
SigStore ADR Proposal because it would have two different configs for
signing and verifications but I think having one type that can be used
with different fields would also work, and its significantly easier to
implement here_

#### Why this change is needed

- Expands the plugin framework capabilities to include signing handlers
for enhanced security and verification processes.

part of adopting the RSA handler implemented for
https://github.com/open-component-model/ocm-project/issues/648 in
https://github.com/open-component-model/open-component-model/pull/859.
It would be registered as an internal handler via
`RegisterInternalComponentSignatureHandler` in the CLI

---------

Signed-off-by: Jakob Möller <jakob.moeller@sap.com>