fix(fetch): streamline blob retrieval by always referencing the digest instead of the annotations (#986)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it

we can significantly streamline reading of local blobs by using the
local reference from the access instead of its identity annotation.

the added benefit of this is that we will now be able to read OCM OCI
descriptors even if they do not have these annotations set. This is
possible because we now can directly use the reference digest. This
makes us compatible with any OCM component version created without these
(not standard, but created as such by old versions of CLI and other
tooling)

#### Which issue(s) this PR fixes
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->

fixes two broken test cases as well as reading without digest (the fact
that the test was broken shows that the digest lookup works because the
CV in the test had wrong digests that were before working just because
the digest was ignored)

Signed-off-by: Jakob Möller <jakob.moeller@sap.com>
Co-authored-by: Gergely Brautigam <gergely.brautigam@sap.com>