fix(cli): multi ctf repo race (#965)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it
This PR fixes is supposed to fix the race conditions occuring when
multiple ctf instances operate on the same ctf archive. The actual fix
is provided in
https://github.com/open-component-model/open-component-model/pull/964
and
https://github.com/open-component-model/open-component-model/pull/963.
Here, a the archive instance and its lock are shared between multiple
ctf repositories operating on the archive.

#### Which issue(s) this PR fixes
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->
Contributes to
https://github.com/open-component-model/ocm-project/issues/564

---------

Signed-off-by: Fabian Burth <fabian.burth@sap.com>