chore(cli): adjust cli to dag refactor (#1032)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it
- Adjust the rendering logic to the refactored dag.
- Adjusts the `get cv` command to the refactored dag and adjusted
rendering logic

**Special Notes for the Reviewer:**
I didn't want this adjustment to end up in an entire rewrite, also
because it would be harder to review. But we should probably discuss
whether the rendering logic should be working with generics as well
instead of directly accessing the attributes of the graph.

#### Which issue(s) this PR fixes
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->
Contributes to
https://github.com/open-component-model/open-component-model/issues/903

---------

Signed-off-by: Fabian Burth <fabian.burth@sap.com>